chore: All tests run  now
