+++ src/Http/Controllers/Auth/TwoFactorController.php
@@ -2,8 +2,9 @@
 
 namespace Whilesmart\UserAuthentication\Http\Controllers\Auth;
 
-use App\Http\Controllers\Controller;
+// use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
+use Illuminate\Routing\Controller;
 use Illuminate\Support\Facades\Auth;
 
 class TwoFactorController extends Controller
